Move show to gui ns
The ns dependency on mikera.gui leads to a dependency on
sun.java2d.opengl.OGLRenderQueue which spawns a thread during its static
creation. This is unneeded overhead for anyone not using the show fn and
it interferes with native-image compilation with graal. This refactor
makes GUI, and thus OpenGL, support opt-in.